refactor(nexus-sandbox): split lifecycle.ts to satisfy complexity ratchet
Lifecycle file grew to 619 lines after the adversarial-review hardening,
exceeding the soft 400-line limit and pushing the repo over the
check:complexity ratchet (729 vs 728 budget). Extract two focused
helper modules:

- port-lock.ts: per-port advisory lock + stale-takeover sentinel
- process-tree.ts: lsof-based listener-ownership verification + process
  tree walking (descendants snapshot, parent-pid lookup, liveness check)

lifecycle.ts now hosts only the spawn/shutdown orchestration (336 lines)
plus a small `verifyOwnershipAfterHealth` helper extracted to keep
startSandbox under the 50-line function limit.

No behavioral change.
